Added some fixes and removed comments

+++ Student Management System/src/main/java/org/example/studentmanagementsystem/model/entities/Student.java
@@ -43,7 +43,7 @@ public class Student extends User {
         for (Grade grade : grades) {
             sum += grade.getGrade();
         }
-        return (Double) (sum / grades.size());
+        return sum / grades.size();
     }
 
 }
